requirements.txt
# requirements.txt
langchain==0.1.0
pymilvus==2.3.3
streamlit==1.30.0
sentence-transformers==2.2.2
pypdf==3.17.1
python-dotenv==1.0.0
huggingface-hub==0.20.1
torch==2.1.2
